+++ test/main.test.ts
@@ -7631,7 +7631,7 @@ ExternalRef: PACKAGE-MANAGER purl pkg:npm/noassertion-spdx-tag-value-child@1.0.0
         [
           "packages:",
           "- completed:",
-          "    hackage: risk-haskell-1.2.3@sha256:abc,1234",
+          `    hackage: risk-haskell-1.2.3@sha256:${"a".repeat(64)},1234`,
           "  original:",
           "    hackage: risk-haskell-1.2.3"
         ].join("\n"),